% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxel_functions.R
\name{estimate_max_returns}
\alias{estimate_max_returns}
\title{Estimate the potential maximum number of returns for voxels}
\usage{
estimate_max_returns(scandata, xysize, zsize, numcols, zmax = 35.25,
  zmin = NULL)
}
\arguments{
\item{scandata}{A three column matrix of XYZ coordinates for return positions
relative to the scanner.}

\item{xysize}{The horizontal size (i.e. width in the X-Y plane) of voxels in
metres.}

\item{zsize}{The vertical size (height) of voxels in metres.}

\item{numcols}{The integer number of voxel colums in the X and Y direction.
This should be an odd value so that the scanner lies in the centre column.
If an even value is provided it will be adjusted to the next odd value.}

\item{zmax}{The maximum height of the scan (relative to the scanner) in
metres. Default value is 35.25 as used by Ashcroft et al. This will usually
be a positive value (uppermost voxel above scanner) but zero and negative
values are also valid.}

\item{zmin}{The minimum height of the scan (relative to the scanner) in
metres. Must be less than \code{zmax}. Defaults to \code{-zmax} if zmax is
positive. If \code{zmax} is zero or negative an explicit value must be
provided.}
}
\value{
A matrix with voxel horizontal position as rows (X varies fastest,
  then Y) and voxel height as columns.
}
\description{
This function takes data from a control scan that was performed in a closed
room (i.e. no gaps or reflective surfaces) where all laser pulses are
expected to provide a return.
}
