% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxel_functions.R
\name{get_max_returns}
\alias{get_max_returns}
\title{Estimate the potential maximum number of returns for voxels}
\usage{
get_max_returns(scandata, xysize, zsize, numcols, zmax = 35.25, zmin = NULL)
}
\arguments{
\item{scandata}{Either a three column matrix or data frame of XYZ coordinates
for return positions relative to the scanner; or a \code{LAS} object as
returned by \code{lidR::readLAS}.}

\item{xysize}{The horizontal size (i.e. width in the X-Y plane) of voxels in
metres.}

\item{zsize}{The vertical size (height) of voxels in metres.}

\item{numcols}{The integer number of voxel colums in the X and Y direction.
Ashcroft et al. recommend that this be an odd value so that the scanner
lies in the centre column, but this is not enforced here.}

\item{zmax}{The maximum height of the scan (relative to the scanner) in
metres. Default value is 35.25 as used by Ashcroft et al. This will usually
be a positive value (uppermost voxel above scanner) but zero and negative
values are also valid.}

\item{zmin}{The minimum height of the scan (relative to the scanner) in
metres. Must be less than \code{zmax}. Defaults to minus \code{zmax} if
zmax is positive. If \code{zmax} is zero or negative an explicit value must
be provided.}
}
\value{
A named list (object of class \code{TLSNullModel}) with the following
  elements:
  \describe{
  \item{returns}{An array (dimensions: Yrow, Xcol, Zlayer) of maximum
    expected number of rays passing through each voxel.}
  \item{xysize}{The horizontal voxel width.}
  \item{xlims}{A vector of min and max values for X and Y ordinates.}
  \item{zsize}{The vertical voxel width.}
  \item{zlims}{A vector or min and max values for Z (height).}
  \item{numcols}{Number of voxels in each horizontal dimension.}
  \item{numheights}{Number of voxels in vertical dimension.}
  }
}
\description{
This function takes data from a control scan that was performed in a closed
room (i.e. no gaps or reflective surfaces) where all laser pulses are
expected to provide a return.
}
\examples{
\dontrun{
# Estimate returns for regular cubic voxels with 0.25m side length,
# for a scan block with horizontal dimensions 60x60 voxels
refdata <- get_max_returns(control.scan,
                           xysize = 0.25, zsize = 0.25, numcols = 60,
                           zmax = 100)
}

}
