% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxel_functions.R
\name{get_prop_reflected}
\alias{get_prop_reflected}
\title{Main function to process scans}
\usage{
get_prop_reflected(scandata, refdata, ground)
}
\arguments{
\item{scandata}{A three column matrix of XYZ coordinates for return positions
relative to the scanner.}

\item{refdata}{A named list, as returned by the functcion
\code{get_max_returns()}, containing an array of the estimated maximum
number of returns for each voxel plus the voxel and scan dimensions.}

\item{ground}{A square matrix where values are ground elevation for each
voxel. The dimensions of the matrix must match the value of element
\code{numcols} in \code{refdata}.}
}
\value{
A named list with the following elements:
  \describe{
  \item{preflect}{An array (dimensions: X, Y, Z) of the proportion of rays
    reflected for each voxel.}
  \item{xysize}{The horizontal voxel width.}
  \item{xlims}{A vector of min and max values for X and Y ordinates.}
  \item{zsize}{The vertical voxel width.}
  \item{zlims}{A vector or min and max values for Z (height).}
  \item{numcols}{NUmber of voxels in each horizontal dimension.}
  \item{numheights}{Number of voxels in vertical dimension.}
  }
}
\description{
This function takes scan data for a vegetation site, together with a reference
object giving the maximum expected number of returns per voxel, and calculates
the proportion of rays reflected for each voxel. The reference object is
generated using the function \code{get_max_returns()}.
}
