% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxel_functions.R
\name{nibble_matrix}
\alias{nibble_matrix}
\title{Iteratively replace NA values in a numeric matrix}
\usage{
nibble_matrix(M)
}
\arguments{
\item{M}{A numeric matrix possibly containing missing values}
}
\value{
A numeric matrix with missing values replaced. If all values were
  missing in the input matrix, a warning message is issues and the matrix is
  returned unchanged.
}
\description{
This function iteratively replaces missing values in a numeric matrix with
the mean of neighbouring cell values. Both orthogonal and diagonal neighbours
are included when calculating the mean value. The function is used by
\code{get_prop_reflected} to remove missing values from ground elevation
matrices.
}
