% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_helpers.R
\name{res_to_brick}
\alias{res_to_brick}
\title{Create a raster brick from \code{get_prop_reflected} results}
\usage{
res_to_brick(
  res,
  elem = c("preflected", "nreflected", "nthrough", "nblocked", "lower", "upper")
)
}
\arguments{
\item{res}{A \code{TLSResult} object as returned by \code{get_prop_reflected}.}

\item{elem}{The result element to convert to a \code{RasterBrick}. One of:
'preflected' (default), 'nreflected', 'nthrough', 'nblocked', 'lower', 'upper'.
An abbreviated name will work if unique. The 'lower' and 'upper' options only
apply if the result object contains arrays for lower and upper bounds of the
proportion of reflected rays.}
}
\description{
This function takes a result list object returned by
\code{\link{get_prop_reflected}} and creates a \code{RasterBrick} object
where the cell values in each layer represent the proportion of rays returned
for each voxel column.
}
