% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxel_functions.R
\name{get_prop_reflected}
\alias{get_prop_reflected}
\title{Main function to process scans}
\usage{
get_prop_reflected(scandata, refdata, ground)
}
\arguments{
\item{scandata}{A three column matrix of XYZ coordinates for return positions
relative to the scanner.}

\item{refdata}{A named list, as returned by the functcion
\code{get_max_returns()}, containing a matrix of the estimated maximum
number of returns for each voxel plus the voxel and scan dimensions.}

\item{ground}{A square matrix where values are ground elevation for each
voxel. The dimensions of the matrix must match the value of element
\code{numcols} in \code{refdata}.}
}
\value{
A matrix where cell values are proportion of rays reflected for each voxel
  based on the reference data for the potential maximum number.
}
\description{
This function takes scan data for a vegetation site, together with a matrix of
max returns per voxel generated using the function \code{get_max_returns()},
and calculates something...
}
